Split out binary Exif tests
It appears binary Exif values don't currently work (either a libmmal or
camera firmware bug - picamera is passing the values down to libmmal but
they seem to get ignored). Split out the tests for binary Exif values
and marked them as expected-failure and added a docs note about the bug.