Common superclass for image encoders
Replaced _PiStillEncoder and _PiSequenceEncoder with a common
superclass: _PiImageEncoder which has _PiOneImageEncoder (for capture
and capture_continuous) and _PiMultiImageEncoder (for capture_sequence).
Tweaked the encoders so that they're the ones that do the test for
closed cameras and pre-existing encoders - no point duplicating the code
in all the different capture methods.