Added a new recipes chapter
This chapter should probably be fleshed out into several chapters in a
future revision, but for now it provides a good overview of the more
complex bits of functionality in the library (generators for
capture_sequence, conversion of YUV data to RGB, capturing to in-memory
streams, etc. etc). This closes #13