Fix #22
Ensure camera's preview port is permanently connected to some component
(preview or null sink) to force it to measure exposure continually and
prevent captured images gradually fading to black.

Moved from having two connection references to a single connection that
switches between preview and null-sink. Now that a preview is always
(effectively) running and we disable the preview connection where
necessary, there's little point raising an exception if the user
attempts to adjust certain properties. The preview may flick off and on
if it's running but that's not a huge deal.