Fix #24
Crop's getter now works. Unfortunately the resetting of non-rectangular
zooms doesn't seem to but I suspect that's a firmware issue (what sort
of phone is ever going to let you do a non-rectangular zoom...) so
there's not much we can do about it unless we fancy deprecating crop and
replacing it with pan & zoom attributes?