Fix #26
Implement video-port-based capture while recording. This requires the
introduction of a splitter on the video port, rejigging the encoders to
accept an arbitrary input port (it's now either a camera port or a
splitter port, and moving the camera capture-enable code to the camera
class as the encoders can no longer assume their input port is actually
attached to the camera).

In the course of enabling this, some tidy-up of the image encoder
classes has also occurred with the result that raw video-port captures
are also now supported (tests have yet to be added for this).

Make splitter ports explicit