Fix #27
Implemented segmentation of video files via a new split_recording
method. This is a considerably simpler API than that proposed in the
ticket, although it does have the (minor) disadvantage that size-based
segmentation won't be as precise as it would with, say, a callback based
interface.

Also fixed a nasty bug in the MMAL translation (an off-by-1 error due to
a missing entry in the video encoder parameters), which in turn fixed
the setting of immutable input. Finally, tidied up the encoder
configuration code to use explicitly named options (it's a bit more
robust).