Fix #35
Remove GPIO cleanup. It's not necessary for safety as GPIO5 isn't
exposed on the board and breaks people's code that are using GPIO after
camera close. Made LED initialization lazy (on first use) so we also
don't break people's code if they're using GPIO in BOARD mode (as long
as they don't try and use the led attribute).

Also added recipe for the led attribute.