Fix #32
Changed raw captures to have format specified directly on each capture
method, and used the resizer's output format to handle RGB conversion.
Unfortunately RGB encoding isn't accepted by the resizer, but RGBA is so
some manual code is added to the raw encoder class to handle stripping
the unused alpha bytes.

As the raw encoder stuff is common to both the one and multi image
encoders, and as it's now rather more complicated, it also made sense to
consolidate the code in a mixin class.