Fix #59
This is a bit of a horrid hack, but it seems there's no other solution.
When still capture is requested without video-port or resizer usage, the
camera is reconfigured to use opaque encoding on the still port.
Otherwise, the camera is reconfigured to use YUV 4:2:0 on the still
port.

Reconfiguration only happens if required when capture methods are used,
but the recording methods force the still port encoding to YUV 4:2:0 to
ensure that if capture-while-record (from the video port) is utilized it
doesn't crash the recording.