Fix #83 and re-work raw captures
Re-wrote a lot of the code surrounding raw (YUV/RGB/RGBA/etc.) captures
partially to fix #83 and partially to bring picamera much closer to
raspiyuv in behaviour. Resolution values in camera and encoder
configurations are now aligned appropriately (to 32 for width, and 16
for height).

More importantly, picamera only uses a resizer to perform raw
conversions when absolutely necessary (i.e. when video-port based
captures involving the splitter are requested, or when a resize is
explicitly requested). Otherwise, it manipulates the encoding of the
still port and dumps data directly from the port output. This also
solves the long-standing issue with RGB captures being slower than RGBA
(although only for unresized still-port-based captures).

The test suite has been updated to remove the relevant xfails as well.