Tweak some boundaries in the tests
Got a couple of failures in the last test run due to some of the limits
being a bit strict in some of the weirder cases that we test (e.g. 5fps
failed on a split because it waited so long for an I-frame to come along
that the timing broke the 20% margin I'd allowed).