Add mmalobj layer
This is a major re-factoring of picamera's guts. It introduces the
"mmalobj" module which is a thin layer on top of mmal intended to make
it a bit easier to work with from Python (but nowhere near as high-level
as picamera's own API).

The picamera API itself is then re-written to be based solely on the
mmalobj layer which it relies upon for things like port format
negotiation, automatic pool construction, buffer releasing, etc.