Several Improvements/Fixes Listed in the Long Descrit
Edits made with the help of @maxwelljg

Changed flush() so that it outputs a 2D array which is used by most entities interested in raw data.  Also, removed the np.delete() function from flush() and instead built a new array from the data, cutting 1.5 seconds off the raw image capture time.

Edited demosaic() to work with the 2D array.  Also changed the bayer pattern to the correct pattern.  Previously, the flush and demosaic functions assumed the bayer pattern of the picamera sensor was GBRG when it is actually BGGR.  This resulted in images obtained from the raw data having the incorrect colors.  (Using the demosaic function would still create a correctly colored image since the colors were originally stored in the same incorrect pattern.  Building an image in RGB format from the raw data using the correct pattern (BGGR) would create a miscolored image however.)