Revert "Use the ISP resizer when available"
This reverts commit fdfc5f2b969620f26528e29baa9502498ab38e29. Only
MMALResizer as MMALISPResizer doesn't work with MMALSplitter (connects
but no data flows). This leaves only still port captures capable of
using the ISP resizer, and makes the logic of deciding when to use the
resizer extremely complex (given old firmwares don't have the ISP
resizer and the resizer is sometimes used for unencoded transforms). For
now it's best to leave things using the VPU resizer and revisit this
if/when the firmware gets patched up.