Connect encoder upon construction
Currently, the encoder input isn't committed prior to encoder
configuration which causes an issue with the pool size allocation on the
output port. Connect the encoder upstream (but leave the connection
disabled) prior to configuring it. This fixes an issue with too-small
buffers on the output port of certain encoders.