Re #199
Apparently old firmwares don't support BGRA on renderers (so don't test
it; RGBA is enough to confirm the size calculations). Also, bizarrely,
in Python 2 a memoryview (which is explicitly built on buffer-protocol
objects) is not itself a buffer-protocol object (it is in 3) so just
have buffer_bytes calculate the size in bytes.