Remove custom install routine
Fixes #404.

Not all Pi operating systems report a 'Hardware' line in /proc/cpuinfo.

This would allow the package to be installed on non-Pi systems.  This
isn't really an issue, though: People installing this on non-Pi systems
would simply run into runtime errors.

As I understand it, the setup script should contain as little logic as
possible.  But I guess this is one of those things where people have
opinions one way or the other.

And to quote from the commit message that added this section:

"And this is why you try not to do special cases in your setup.py, kids,
because there's no end once you start!"