doc: Slighly improvement of the documentation with respect the zoom property.
The docstring of the Picamera.zoom property has been improved explaining the
specific behaviour of this property when it is combined with rotation, for example.
Also, it says that it can modify the aspect ratio of the image.
Moreover, in fov.html documentation, it includes the operation order before retrieving
the processed image: first applies the zoom and then rescale the image.