Update mmalobj.py
I think there may just be a bug in the statement at line 464, which is
if not isinstance(buf, memoryview):
in release 1.13. If I remove the "not" and put this instead:
if isinstance(buf, memoryview):
it seems to work fine. It was failing when the condition isinstance(buf, memoryview) was true.