Re-work preservation of timestamps
We don't need all those extra encoder fields, and it'd be nice if the
split_recording method could return the entire frame meta-data rather
than just the timestamp.

This also re-works "None" timestamps in general. We can reasonably
assume the camera's clock never goes backward during a recording so
in the event of a "None" timestamp we should *always* repeat the
last timestamp.