Revise circular buffer
Consolidated most of the _find_* methods into a single generic _find
method. Reduced the time for which the shared lock is held to minimize
the risk of dropped frames when recording is on-going to the circular
buffer. Changed return value of copy_to to return the first and last
frame metadata instead of just the timestamp.

Tweaked the underlying circular buffer logic to drop entire writes
instead of slicing them. The original code acted more like a traditional
ring buffer but the new method is more efficient during writes and the
overall stream limit is still maintained.

Added weakrefs to the sub-ordinate classes used by PiCameraCircularIO
to ensure it can be easily disposed of by the GC.