Re #512
Add colorspace attribute to MMALVideoPort, and a colorspace parameter to
PiCamera.__init__. Color space seems to propogate through the MMAL
pipeline happily once set on the source port. This *should* really be a
recording / capture option but since the splitter is always there it's
better to make it a camera-wide option. Still need to add that as a
PiCamera attribute (like resolution it'll require a call to
_configure_camera to complete).